--- docs/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "mrujs-docs",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "build:src": "NODE_ENV=production BRIDGETOWN_ENV=production bin/bridgetown build",
-    "deploy": "NODE_ENV=production BRIDGETOWN_ENV=production snowpack build",
-    "clean:src": "bundle exec bridgetown clean",
-    "restart": "pnpm clean:src && pnpm start --reload",
-    "start": "NODE_ENV=development BRIDGETOWN_ENV=development snowpack dev"
-  },
-  "dependencies": {
-    "@fullhuman/postcss-purgecss": "^3.1.3",
-    "@snowpack/plugin-postcss": "^1.1.0",
-    "@snowpack/plugin-run-script": "^2.3.0",
-    "@snowpack/plugin-typescript": "^1.2.1",
-    "autoprefixer": "^10.2.1",
-    "bridgetown-quick-search": "1.1.0",
-    "lit": "^2.0.0-rc.2",
-    "lit-html": "^2.1.1",
-    "lunr": "^2.3.9",
-    "morphdom": "^2.6.1",
-    "mrujs": "link:../",
-    "postcss": "^8.2.4",
-    "postcss-cli": "^8.3.1",
-    "prettier": "^2.2.1",
-    "snowpack": "^3.0.11",
-    "stimulus": "^2.0.0",
-    "turbolinks": "^5.2.0"
-  }
-}
