--- docs/postcss.config.js
@@ -1,7 +1,0 @@
-const plugins = {
-  "autoprefixer": {},
-}
-
-module.exports = {
-  plugins
-};
