+++ docs/src/_components/alert.erb
@@ -1,4 +1,4 @@
-<sl-alert type="<%= @type %>" open>
+<sl-alert variant="<%= @type.to_s %>" open>
   <sl-icon slot="icon" name="<%= @icon %>"></sl-icon>
   <%= title %>
   <%= content %>
