+++ docs/src/_partials/_top_nav.erb
@@ -1,4 +1,4 @@
-<div style="display: none;" data-search-target="container" class="search-modal">
+<div hidden data-search-target="container" class="search-modal">
   <bridgetown-search-form class="search-modal__form">
     <sl-icon-button class="search-modal__close" name="x-lg" label="Close"
                     data-action="click->search#hide"></sl-icon-button>
