+++ docs/src/_partials/_version_number.erb
@@ -1,3 +1,3 @@
 <div style="font-size: var(--font-large); margin-left: 0.5rem;">
-  <%= version_number %>
+  <%= "v#{version_number}" %>
 </div>
