+++ docs/src/frontend/javascript/controllers/search_controller.js
@@ -1,0 +1,2 @@
+import { Controller } from "stimulus"
+
