+++ docs/src/frontend/styles/variables.css
