+++ plugins/src/index.ts
@@ -13,4 +13,3 @@ export * from './cableCar'
 export * from './jsErb'
 export * from './mrujsTurbo'
 export * from './shoelace'
-export * from './activeStorage'
